tests for visitors WIP
