+++ src/Nethermind/Nethermind.Blockchain.Test/BlockTreeTests.cs
@@ -638,75 +638,6 @@ namespace Nethermind.Blockchain.Test
         }
 
         [Test]
-        public async Task Can_load_blocks_from_db()
-        {
-            for (int chainLength = 30; chainLength <= 32; chainLength++)
-            {
-                Block genesisBlock = Build.A.Block.Genesis.TestObject;
-
-                MemDb blocksDb = new MemDb();
-                MemDb blockInfosDb = new MemDb();
-                MemDb headersDb = new MemDb();
-
-                BlockTree testTree = Build.A.BlockTree(genesisBlock).OfChainLength(chainLength).TestObject;
-                for (int i = 0; i < testTree.Head.Number + 1; i++)
-                {
-                    Block ithBlock = testTree.FindBlock(i, BlockTreeLookupOptions.None);
-                    blocksDb.Set(ithBlock.Hash, Rlp.Encode(ithBlock).Bytes);
-                    
-                    headersDb.Set(ithBlock.Hash, Rlp.Encode(ithBlock.Header).Bytes);
-
-                    ChainLevelInfo ithLevel = new ChainLevelInfo(true, new BlockInfo[1] {new BlockInfo(ithBlock.Hash, ithBlock.TotalDifficulty.Value) {WasProcessed = true}});
-                    blockInfosDb.Set(i, Rlp.Encode(ithLevel).Bytes);
-                }
-
-                blockInfosDb.Set(Keccak.Zero, genesisBlock.Header.Hash.Bytes);
-                headersDb.Set(genesisBlock.Header.Hash, Rlp.Encode(genesisBlock.Header).Bytes);
-
-                BlockTree blockTree = new BlockTree(blocksDb, headersDb, blockInfosDb, new ChainLevelInfoRepository(blockInfosDb), OlympicSpecProvider.Instance, NullTxPool.Instance, NullBloomStorage.Instance, LimboLogs.Instance);
-
-                DbBlocksLoader loader = new DbBlocksLoader(null, null, null, blockTree, LimboNoErrorLogger.Instance);
-                await blockTree.Accept(loader, CancellationToken.None);
-
-                Assert.AreEqual(testTree.Head.Hash, blockTree.BestSuggestedHeader.Hash, $"head {chainLength}");
-            }
-        }
-
-        [Test]
-        public async Task Can_load_blocks_from_db_odd()
-        {
-            for (int chainLength = 2; chainLength <= 32; chainLength++)
-            {
-                Block genesisBlock = Build.A.Block.Genesis.TestObject;
-
-                MemDb blocksDb = new MemDb();
-                MemDb blockInfosDb = new MemDb();
-                MemDb headersDb = new MemDb();
-
-                BlockTree testTree = Build.A.BlockTree(genesisBlock).OfChainLength(chainLength).TestObject;
-                for (int i = 0; i < testTree.Head.Number + 1; i++)
-                {
-                    Block ithBlock = testTree.FindBlock(i, BlockTreeLookupOptions.None);
-                    blocksDb.Set(ithBlock.Hash, Rlp.Encode(ithBlock).Bytes);
-                    
-                    headersDb.Set(ithBlock.Hash, Rlp.Encode(ithBlock.Header).Bytes);
-
-                    ChainLevelInfo ithLevel = new ChainLevelInfo(true, new BlockInfo[1] {new BlockInfo(ithBlock.Hash, ithBlock.TotalDifficulty.Value)});
-                    blockInfosDb.Set(i, Rlp.Encode(ithLevel).Bytes);
-                }
-
-                blockInfosDb.Set(Keccak.Zero, genesisBlock.Header.Hash.Bytes);
-                headersDb.Set(genesisBlock.Header.Hash, Rlp.Encode(genesisBlock.Header).Bytes);
-
-                BlockTree blockTree = new BlockTree(blocksDb, headersDb, blockInfosDb, new ChainLevelInfoRepository(blockInfosDb), OlympicSpecProvider.Instance, Substitute.For<ITxPool>(), NullBloomStorage.Instance, LimboLogs.Instance);
-                DbBlocksLoader loader = new DbBlocksLoader(null, null, null, blockTree, LimboNoErrorLogger.Instance);
-                await blockTree.Accept(loader, CancellationToken.None);
-
-                Assert.AreEqual(testTree.Head.Hash, blockTree.BestSuggestedHeader.Hash, $"head {chainLength}");
-            }
-        }
-
-        [Test]
         public void Sets_head_block_hash_in_db_on_new_head_block()
         {
             MemDb blocksDb = new MemDb();
@@ -853,8 +784,6 @@ namespace Nethermind.Blockchain.Test
             Assert.AreEqual(block1.Header, tree.BestSuggestedHeader);
         }
 
-        private int _dbLoadTimeout = 5000;
-
         [Test]
         public void When_deleting_invalid_block_deletes_its_descendants()
         {
@@ -885,45 +814,6 @@ namespace Nethermind.Blockchain.Test
             Assert.IsNull(blocksDb.Get(block3.Hash), "block 3");
 
             Assert.NotNull(blockInfosDb.Get(1), "level 1");
-            Assert.IsNull(blockInfosDb.Get(2), "level 2");
-            Assert.IsNull(blockInfosDb.Get(3), "level 3");
-        }
-
-        [Test]
-        public async Task Cleans_invalid_blocks_before_starting_DB_load()
-        {
-            MemDb blocksDb = new MemDb();
-            MemDb blockInfosDb = new MemDb();
-            MemDb headersDb = new MemDb();
-            BlockTree tree = new BlockTree(blocksDb, headersDb, blockInfosDb, new ChainLevelInfoRepository(blockInfosDb), MainnetSpecProvider.Instance, NullTxPool.Instance, NullBloomStorage.Instance, LimboLogs.Instance);
-            Block block0 = Build.A.Block.WithNumber(0).WithDifficulty(1).TestObject;
-            Block block1 = Build.A.Block.WithNumber(1).WithDifficulty(2).WithParent(block0).TestObject;
-            Block block2 = Build.A.Block.WithNumber(2).WithDifficulty(3).WithParent(block1).TestObject;
-            Block block3 = Build.A.Block.WithNumber(3).WithDifficulty(4).WithParent(block2).TestObject;
-
-            tree.SuggestBlock(block0);
-            tree.SuggestBlock(block1);
-            tree.SuggestBlock(block2);
-            tree.SuggestBlock(block3);
-
-            blockInfosDb.Set(BlockTree.DeletePointerAddressInDb, block1.Hash.Bytes);
-
-            CancellationTokenSource tokenSource = new CancellationTokenSource();
-#pragma warning disable 4014
-            Task.Delay(_dbLoadTimeout).ContinueWith(t => tokenSource.Cancel());
-#pragma warning restore 4014
-            DbBlocksLoader loader = new DbBlocksLoader(null, null, null, tree, LimboNoErrorLogger.Instance);
-            await tree.Accept(loader, CancellationToken.None);
-
-            Assert.AreEqual(0L, tree.BestKnownNumber, "best known");
-            Assert.AreEqual(null, tree.Head, "head");
-            Assert.AreEqual(0L, tree.BestSuggestedHeader.Number, "suggested");
-
-            Assert.IsNull(blocksDb.Get(block2.Hash), "block 1");
-            Assert.IsNull(blocksDb.Get(block2.Hash), "block 2");
-            Assert.IsNull(blocksDb.Get(block3.Hash), "block 3");
-
-            Assert.IsNull(blockInfosDb.Get(2), "level 1");
             Assert.IsNull(blockInfosDb.Get(2), "level 2");
             Assert.IsNull(blockInfosDb.Get(3), "level 3");
         }
@@ -975,41 +865,6 @@ namespace Nethermind.Blockchain.Test
         }
 
         [Test]
-        public void When_head_block_is_followed_by_a_block_bodies_gap_it_should_delete_all_levels_after_the_gap_start()
-        {
-            MemDb blocksDb = new MemDb();
-            MemDb blockInfosDb = new MemDb();
-            MemDb headersDb = new MemDb();
-            BlockTree tree = new BlockTree(blocksDb, headersDb, blockInfosDb, new ChainLevelInfoRepository(blockInfosDb), MainnetSpecProvider.Instance, NullTxPool.Instance, NullBloomStorage.Instance, LimboLogs.Instance);
-            Block block0 = Build.A.Block.WithNumber(0).WithDifficulty(1).TestObject;
-            Block block1 = Build.A.Block.WithNumber(1).WithDifficulty(2).WithParent(block0).TestObject;
-            Block block2 = Build.A.Block.WithNumber(2).WithDifficulty(3).WithParent(block1).TestObject;
-            Block block3 = Build.A.Block.WithNumber(3).WithDifficulty(4).WithParent(block2).TestObject;
-            Block block4 = Build.A.Block.WithNumber(4).WithDifficulty(5).WithParent(block3).TestObject;
-            Block block5 = Build.A.Block.WithNumber(5).WithDifficulty(6).WithParent(block4).TestObject;
-
-            tree.SuggestBlock(block0);
-            tree.SuggestBlock(block1);
-            tree.SuggestBlock(block2);
-            tree.SuggestHeader(block3.Header);
-            tree.SuggestHeader(block4.Header);
-            tree.SuggestBlock(block5);
-
-            tree.UpdateMainChain(block2);
-
-            StartupBlockTreeFixer fixer = new StartupBlockTreeFixer(tree, LimboNoErrorLogger.Instance);
-            tree.Accept(fixer, CancellationToken.None);
-
-            Assert.Null(blockInfosDb.Get(3), "level 3");
-            Assert.Null(blockInfosDb.Get(4), "level 4");
-            Assert.Null(blockInfosDb.Get(5), "level 5");
-
-            Assert.AreEqual(2L, tree.BestKnownNumber, "best known");
-            Assert.AreEqual(block2.Header, tree.Head?.Header, "head");
-            Assert.AreEqual(block2.Hash, tree.BestSuggestedHeader.Hash, "suggested");
-        }
-
-        [Test]
         public void When_deleting_invalid_block_does_not_delete_blocks_that_are_not_its_descendants()
         {
             MemDb blocksDb = new MemDb();
@@ -1040,173 +895,6 @@ namespace Nethermind.Blockchain.Test
             Assert.AreEqual(5L, tree.BestKnownNumber, "best known");
             Assert.AreEqual(block5.Header, tree.Head?.Header, "head");
             Assert.AreEqual(block5.Hash, tree.BestSuggestedHeader.Hash, "suggested");
-        }
-
-        [Test]
-        public async Task When_cleaning_descendants_of_invalid_does_not_touch_other_branches()
-        {
-            MemDb blocksDb = new MemDb();
-            MemDb blockInfosDb = new MemDb();
-            MemDb headersDb = new MemDb();
-            BlockTree tree = new BlockTree(blocksDb, headersDb, blockInfosDb, new ChainLevelInfoRepository(blockInfosDb), MainnetSpecProvider.Instance, NullTxPool.Instance, NullBloomStorage.Instance, LimboLogs.Instance);
-            Block block0 = Build.A.Block.WithNumber(0).WithDifficulty(1).TestObject;
-            Block block1 = Build.A.Block.WithNumber(1).WithDifficulty(2).WithParent(block0).TestObject;
-            Block block2 = Build.A.Block.WithNumber(2).WithDifficulty(3).WithParent(block1).TestObject;
-            Block block3 = Build.A.Block.WithNumber(3).WithDifficulty(4).WithParent(block2).TestObject;
-
-            Block block1B = Build.A.Block.WithNumber(1).WithDifficulty(1).WithParent(block0).TestObject;
-            Block block2B = Build.A.Block.WithNumber(2).WithDifficulty(1).WithParent(block1B).TestObject;
-            Block block3B = Build.A.Block.WithNumber(3).WithDifficulty(1).WithParent(block2B).TestObject;
-
-            tree.SuggestBlock(block0);
-            tree.SuggestBlock(block1);
-            tree.SuggestBlock(block2);
-            tree.SuggestBlock(block3);
-
-            tree.SuggestBlock(block1B);
-            tree.SuggestBlock(block2B);
-            tree.SuggestBlock(block3B);
-
-            blockInfosDb.Set(BlockTree.DeletePointerAddressInDb, block1.Hash.Bytes);
-
-            CancellationTokenSource tokenSource = new CancellationTokenSource();
-#pragma warning disable 4014
-            Task.Delay(_dbLoadTimeout).ContinueWith(t => tokenSource.Cancel());
-#pragma warning restore 4014
-            DbBlocksLoader loader = new DbBlocksLoader(null, null, null, tree, LimboNoErrorLogger.Instance);
-            await tree.Accept(loader, CancellationToken.None);
-
-            Assert.AreEqual(3L, tree.BestKnownNumber, "best known");
-            Assert.AreEqual(null, tree.Head, "head");
-            Assert.AreEqual(block3B.Hash, tree.BestSuggestedHeader.Hash, "suggested");
-
-            Assert.IsNull(blocksDb.Get(block1.Hash), "block 1");
-            Assert.IsNull(blocksDb.Get(block2.Hash), "block 2");
-            Assert.IsNull(blocksDb.Get(block3.Hash), "block 3");
-
-            Assert.NotNull(blockInfosDb.Get(1), "level 1");
-            Assert.NotNull(blockInfosDb.Get(2), "level 2");
-            Assert.NotNull(blockInfosDb.Get(3), "level 3");
-        }
-
-        [Test]
-        public async Task Can_load_from_DB_when_there_is_an_invalid_block_in_DB_and_a_valid_branch()
-        {
-            MemDb blocksDb = new MemDb();
-            MemDb blockInfosDb = new MemDb();
-            MemDb headersDb = new MemDb();
-            BlockTree tree1 = new BlockTree(blocksDb, headersDb, blockInfosDb, new ChainLevelInfoRepository(blockInfosDb), MainnetSpecProvider.Instance, NullTxPool.Instance, NullBloomStorage.Instance, LimboLogs.Instance);
-
-            Block block0 = Build.A.Block.WithNumber(0).WithDifficulty(1).TestObject;
-            Block block1 = Build.A.Block.WithNumber(1).WithDifficulty(2).WithParent(block0).TestObject;
-            Block block2 = Build.A.Block.WithNumber(2).WithDifficulty(3).WithParent(block1).TestObject;
-            Block block3 = Build.A.Block.WithNumber(3).WithDifficulty(4).WithParent(block2).TestObject;
-
-            Block block1B = Build.A.Block.WithNumber(1).WithDifficulty(1).WithParent(block0).TestObject;
-            Block block2B = Build.A.Block.WithNumber(2).WithDifficulty(1).WithParent(block1B).TestObject;
-            Block block3B = Build.A.Block.WithNumber(3).WithDifficulty(1).WithParent(block2B).TestObject;
-
-            tree1.SuggestBlock(block0);
-            tree1.SuggestBlock(block1);
-            tree1.SuggestBlock(block2);
-            tree1.SuggestBlock(block3);
-
-            tree1.SuggestBlock(block1B);
-            tree1.SuggestBlock(block2B);
-            tree1.SuggestBlock(block3B);
-
-            tree1.UpdateMainChain(block0);
-
-            BlockTree tree2 = new BlockTree(blocksDb, headersDb, blockInfosDb, new ChainLevelInfoRepository(blockInfosDb), MainnetSpecProvider.Instance, NullTxPool.Instance, NullBloomStorage.Instance, LimboLogs.Instance);
-
-            CancellationTokenSource tokenSource = new CancellationTokenSource();
-#pragma warning disable 4014
-            Task.Delay(_dbLoadTimeout).ContinueWith(t => tokenSource.Cancel());
-#pragma warning restore 4014
-
-            tree2.NewBestSuggestedBlock += (sender, args) =>
-            {
-                if (args.Block.Hash == block1.Hash)
-                {
-                    tree2.DeleteInvalidBlock(args.Block);
-                }
-                else
-                {
-                    tree2.UpdateMainChain(args.Block);
-                }
-            };
-
-            DbBlocksLoader loader = new DbBlocksLoader(null, 1, null, tree2, LimboNoErrorLogger.Instance);
-            await tree2.Accept(loader, tokenSource.Token);
-
-            Assert.AreEqual(3L, tree2.BestKnownNumber, "best known");
-            Assert.AreEqual(block3B.Hash, tree2.Head.Hash, "head");
-            Assert.AreEqual(block3B.Hash, tree2.BestSuggestedHeader.Hash, "suggested");
-
-            Assert.IsNull(blocksDb.Get(block1.Hash), "block 1");
-            Assert.IsNull(blocksDb.Get(block2.Hash), "block 2");
-            Assert.IsNull(blocksDb.Get(block3.Hash), "block 3");
-
-            Assert.NotNull(blockInfosDb.Get(1), "level 1");
-            Assert.NotNull(blockInfosDb.Get(2), "level 2");
-            Assert.NotNull(blockInfosDb.Get(3), "level 3");
-        }
-
-        [Test]
-        public async Task Can_load_from_DB_when_there_is_only_an_invalid_chain_in_DB()
-        {
-            MemDb blocksDb = new MemDb();
-            MemDb blockInfosDb = new MemDb();
-            MemDb headersDb = new MemDb();
-            BlockTree tree1 = new BlockTree(blocksDb, headersDb, blockInfosDb, new ChainLevelInfoRepository(blockInfosDb), MainnetSpecProvider.Instance, NullTxPool.Instance, NullBloomStorage.Instance, LimboLogs.Instance);
-
-            Block block0 = Build.A.Block.WithNumber(0).WithDifficulty(1).TestObject;
-            Block block1 = Build.A.Block.WithNumber(1).WithDifficulty(2).WithParent(block0).TestObject;
-            Block block2 = Build.A.Block.WithNumber(2).WithDifficulty(3).WithParent(block1).TestObject;
-            Block block3 = Build.A.Block.WithNumber(3).WithDifficulty(4).WithParent(block2).TestObject;
-
-            tree1.SuggestBlock(block0);
-            tree1.SuggestBlock(block1);
-            tree1.SuggestBlock(block2);
-            tree1.SuggestBlock(block3);
-
-            tree1.UpdateMainChain(block0);
-
-            BlockTree tree2 = new BlockTree(blocksDb, headersDb, blockInfosDb, new ChainLevelInfoRepository(blockInfosDb), MainnetSpecProvider.Instance, NullTxPool.Instance, NullBloomStorage.Instance, LimboLogs.Instance);
-
-            CancellationTokenSource tokenSource = new CancellationTokenSource();
-#pragma warning disable 4014
-            Task.Delay(_dbLoadTimeout).ContinueWith(t => tokenSource.Cancel());
-#pragma warning restore 4014
-
-            tree2.NewBestSuggestedBlock += (sender, args) =>
-            {
-                if (args.Block.Hash == block1.Hash)
-                {
-                    tree2.DeleteInvalidBlock(args.Block);
-                }
-                else
-                {
-                    tree2.UpdateMainChain(args.Block);
-                }
-            };
-
-            DbBlocksLoader loader = new DbBlocksLoader(null, 1, null, tree2, LimboNoErrorLogger.Instance);
-            await tree2.Accept(loader, tokenSource.Token);
-
-            /* note the block tree historically loads one less block than it could */
-
-            Assert.AreEqual(0L, tree2.BestKnownNumber, "best known");
-            Assert.AreEqual(block0.Hash, tree2.Head.Hash, "head");
-            Assert.AreEqual(block0.Hash, tree2.BestSuggestedHeader.Hash, "suggested");
-
-            Assert.IsNull(blocksDb.Get(block1.Hash), "block 1");
-            Assert.IsNull(blocksDb.Get(block2.Hash), "block 2");
-            Assert.IsNull(blocksDb.Get(block3.Hash), "block 3");
-
-            Assert.IsNull(blockInfosDb.Get(1), "level 1");
-            Assert.IsNull(blockInfosDb.Get(2), "level 2");
-            Assert.IsNull(blockInfosDb.Get(3), "level 3");
         }
 
         [Test, TestCaseSource("SourceOfBSearchTestCases")]
